edits for past_to_func
